// Patientor/patientor-frontend/src/types.ts
export interface Diagnosis {
  find(arg0: (diagnose: Diagnosis) => boolean): any;
  code: string;
  name: string;
  latin?: string;
}

export type FinderFunction = (a: string) => string;

export enum Gender {
  Male = "male",
  Female = "female",
  Other = "other"
}

export interface Patient {
  id: string;
  name: string;
  occupation: string;
  gender: Gender;
  ssn?: string;
  dateOfBirth?: string;
  entries: Entry[];
}

export type Entry = HealthCheckEntry | OccupationalHealthcareEntry | HospitalEntry;

interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis['code']>;
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

interface OccupationalHealthcareEntry extends BaseEntry {
  type: 'OccupationalHealthcare',
  employerName: string,
  sickLeave?: { startDate: string, endDate: string }
}

interface HospitalEntry extends BaseEntry {
  type: 'Hospital',
  discharge: { date: string, criteria: string }
}